Fix: Allow reads consisting only of insertion + Fix CI

+++ src/genie/core/classifier-regroup.cc
@@ -212,7 +212,7 @@ record::Chunk ClassifierRegroup::getChunk() {
     static int i = 0;
     std::ofstream tmpOut("AU_" + std::to_string(i++) + ".mgrec");
     util::BitWriter bw(&tmpOut);
-    for (const auto &r : ret.getData()) {
+    for (const auto& r : ret.getData()) {
         r.write(bw);
     }
 #endif
@@ -266,14 +266,15 @@ void ClassifierRegroup::add(record::Chunk&& c) {
                           dynamic_cast<const core::record::alignment_split::SameRec&>(
                               *r.getAlignments().front().getAlignmentSplits().front())
                               .getDelta();
-                    end = std::max(end, r.getAlignments().front().getPosition() + r.getMappedLength(0, 0));
+                    end = std::max(static_cast<uint64_t>(end),
+                                   r.getAlignments().front().getPosition() + r.getMappedLength(0, 0));
                 }
                 record_reference = this->refMgr->load(refMgr->ID2Ref(r.getAlignmentSharedData().getSeqID()),
                                                       r.getAlignments().front().getPosition(), end);
             }
         }
 
-        if(currentChunks[refBased][paired][(uint8_t)classtype - 1].getData().empty()) {
+        if (currentChunks[refBased][paired][(uint8_t)classtype - 1].getData().empty()) {
             currentChunks[refBased][paired][(uint8_t)classtype - 1].getRef() = record_reference;
         } else {
             currentChunks[refBased][paired][(uint8_t)classtype - 1].getRef().merge(record_reference);
